added profile page very early in dev

+++ server/schemas/typeDefs.js
@@ -3,7 +3,7 @@ const typeDefs = `
     _id: ID!
     username: String!
     email: String!
-    password: String!
+    password: String! 
   }
   type Bet {
     _id: ID!
@@ -29,7 +29,7 @@ type Query {
   type Mutation {
     addUser(username: String!, email: String!, password: String!): Auth
     login(email: String!, password: String!): Auth
-    addBet(betType: String!, matchup: String!, spread: Int, winner: String, total: Int, endTime: Int!, betStatus: String!, units: Int!): Bet
+    addBet(betType: String!, matchup: String!, spread: Float, winner: String, total: Float, endTime: Int!, betStatus: String!, units: Int!): Bet
 
   }
 `;
